Fixed issue with decimal points

+++ src/com/org/gnos/db/model/FixedOpexCost.java
@@ -1,10 +1,11 @@
 package com.org.gnos.db.model;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 
 public class FixedOpexCost {
 	protected int id;
-	protected HashMap<Integer, Float> costData;
+	protected HashMap<Integer, BigDecimal> costData;
 	protected int scenarioId;
 	
 	public int getId() {
@@ -14,11 +15,11 @@ public class FixedOpexCost {
 		this.id = id;
 	}
 	
-	public HashMap<Integer, Float> getCostData() {
+	public HashMap<Integer, BigDecimal> getCostData() {
 		return costData;
 	}
 
-	public void setCostData(HashMap<Integer, Float> costData) {
+	public void setCostData(HashMap<Integer, BigDecimal> costData) {
 		this.costData = costData;
 	}
